General status report
Have been broadly struggling with ember for most of the afternoon. Unsure
of why, but I cannot properly wire the authentication suite GA built for
us into our master, forcing me to work on this branch. Concluded that one
of the inbuilt statements I replaced w/'medici' is the source of the problem,
but unsure of which one exactly.

Not a crippling issue, as I can simply build in this repo and, if necessary,
use it as my deployed front end. Would rather find the resolution, but
current situation is not crippling.

Currently working on the creation of an account. Have both route and component
defined. My templates are working as desired. Next step is the routing process.
Will also need to ensure refinements as I build out.