Think I've isolated the problem
It looks a lot like the data I need to work with is not sticking to my
local database, making it difficult to have the two interface.

Now that I've isolated the problem, I actually need to research how to
solve the damn thing.

In the meantime, building out visual component with dummy data, should
make it easier to plug working back end connection into when the time
comes.