Inadvertently solved account creation
As it turns out, the problem was that I was passing the data incorrectly
in the create account form template. Altering the value types of the
submit solved this problem.

Have made little progress, however, in solving the problem of displaying
my account. Now that there is data in the store, however, it seems likely
that this problem is solvable. Will resolve this, then consider next steps.