Three problems until done with user account functionality
-first, need to figure out how to pass data returned from the store to
the handlebars
-second, need to remove the hardcoding in my route: need to find a way
to pass modular code to this route
-third, need to determine how the routing of the my-account component
will work

Funnily enough, I think it is reasonable to tackle these in descending
order. First, wire up user facing route logic, then use that logic to pass
the proper value to the route, then apply the data to the hbs files.