Working on donations
Currently having some difficulty getting the adapter method I wrote to
fire in the route. Can't test it's viability without doing so first.

Found good SO answer on internet, may work on implementing its functionality

Worried about the fact that donate is not actually really a patch request
anymore; it's probably closer to a get, but it doesn't really interact
w/any of the data in my account table. It was largely placed there due to
legacy code. In a future refactor, donations will get their own table
and this will be a post method. Need to make do w/what I have due to time
constraints, but this is worth bearing in mind.