Some progress on getting data down from the server
Finally able to read into my nested component with a route from the server
seems as though the problem I was having can be attributed to a mismatching
of which identifier belongs on which side of the equals sign.

Still a few things to do. First, needs to be routed properly. I have
it sitting on the accounts route at the moment, which looks like it will
be a fairly busy area. Need to have my-account at its own route.

Second, I have the account getting handled with hard coding. Still not
entirely sure what variables I can use to pull the account linked w/the
current user, but those structures will be the key.