Planning commit
So, I am currently able to ping my server and show the user the current
data at the correct route. There are a few caveats, however, which makes
it impossible to say this is truly done.

First and most important, I need to figure out how to soft-code my server
call. As it stands, I have no idea how to this, only some vague idea that
it revolves around using dynamic routes.

Second is a stylistic problem: namely that the view at both /accounts and
/my-account are the same. I need to find a way to modularize this.

Once both of these are done, I will be in a decent place, able to have a
user create their account and then direct them to a page which shows them
their information.