Broad status update
AS things stand, I was able to get a 1:1 w/Toni. We mostly went over
my issues in the process of manifesting my-account. Concluded that using
simple linkages would be the best way to actually manifest the data. In
practice, this has proven difficult, as my linkages are not really
manifesting like I want them too; once again suffering a knowledge gap here,
need more responses from the issue queue.

Broadly, was able to make progress and actually call on the current user.
This will prove vital for showing the actual user's account when they call
the my-account template. Unfortunately, this requires me to create a relation
between user and account, which has proven extremely finicky.

Regarding my adapter problem, little progress has been made. I desperately
need a 1:1 if I am going to grasp this concept. I have some adapters written,
but it is still entirely unclear to me how the adapter is going to be called.
The longer it takes for me to clear this up, the more inclined to panic
I become.