Able to bind a user to the account
Finally able to bind an account to a user, a problem solved by creating
a back end serializer which attatches the user as an attribute. Next
step is finding a way to leverage this new information. I am not sure
how I can go about using the account ID to return the correct user id. From
there, I will be able to solve both of the major problems which have been
holding me back.