Testing donate method
Have to hardcode the account being targeted at the moment, but finally
able to run tests on the donate method. Currently being thrown a 500
error, meaning I am pinging the server correctly, but passing bad data.
May be fairly close to getting leg one to work. This leg is a bit easier,
as I imagine it will require little reconfiguration on the back end.