Source of 500 error isolated, can begin to be addressed
My front end testing suite is able to ping the donate method at the server,
but is currently bouncing off because it is not sending the param information
as it is supposed to. Very close, pretty much the only thing in my way
rn is figuring out how to package the params sent to the router and send
them as params to my controller.