Leg two working
Able to send bitcoin to the client account with no friction. Same problem
re; hardcoded identity, but this can largely be solved. More important is
that the firing of the donation is entirely bound, leaving little room
for messiness. With donate functionality largely working, obvious next
step is to figure out how this functionality can be integrated into the
broader app. This is a simple test harness, the real challenge starts now
especially worth noting is that data binding could prove problematic. Figuring
out how to pass the clientId to the route is going to be the hinge point
for donation integration.