Breakthrough made
Finally able to display all items from my back end. This was partially
solved by reconfiguring my broken index method on the back end. From there,
reconfigured template route in such a manner that it was able to pass
data down to the display. Next step is to use an if statement in the
calling of the component to throw out all accounts that do not have the
editable tag (used to discern when an account belongs to an authenticated)
user.