Functionality finally working as intended
Also have everything largely wired together as intended.

Not entirely as I wanted, as I do not have a replicable solution for
securing the account of the current user. Will need to figure out how
to stash the data I want to use in the appropriate handlebars, then send
it upwards. Also, it is entirely possible I will want to build out a service
to deal with this. This would allow me to cache my current user and call
it much like I would w/auth.