Refactoring how to search for an account
Ripped out find-account component in exchange for a broader search accounts
route/template approach. End goal is for the user to search for an account
by username, then use the result of this search to redirect the user to
the profile of the searched user. Currently have all templates properly
wired, next step is setting up component.