Able to get actions to properly bubble up from components
Turns out my error was in a failure to grasp how the template side of a
component plays into the transfer of actions to the route. A rectification
of this allows me to hit the route.

Next step is to reconfigure the model hook for my current user to have it
limited to only a single user, rather than a massive and largely unusable
array. Such restriction of data scope will be vital in specifying targets
for both donations and targeted crud actions.