Model hook reconfigured to only pass a single item to the template
However, it is still returning a huge amount of data, which could prove
highly problematic in the passing of data to the donate and update method.

There may be a way to further limit scope, but I think that would require
the use of a dynamic route, which would require a total reconfiguration
of my current file structure. Given the point I am at, this is an difficult
objective.