Update functionality working as intended
Able to update the user name of the current user's account. Getting a
far better understanding of the promise structure and how it relates
to ember. Next step is to delete account. This is a far trickier task,
as deleting an account requires the deletion of the account, then the
deletion of the user, then signing out. Worried about room for error.

Deploying what I have to master.