Donate reconstruction overview
I am attempting to reconfigure the donate method to make it as seamless
as possible. Now, the donate form sits in the view-account template and
routes up to the view-accounts route. This is where the donate method
will be called. What is critical here is that I find a way to pass the
account ids of the current user and the target of the donation up to the
route file, where they will then be passed into the custom adapter methods.

In practice, this means bundling them into the params object at some
level. This is likely easier said than done; I have really only ever
done db queries in the routes, never in the components. What I wonder
is whether or not it is possible to stash the values of both the client
and the patron in the button, which would allow the submittal to pass them
up as values. Certainly worth a shot.