Added module functionality to account create form
The form for creating a client account can now be hidden and shown with
a button. This button also prevents the user from passing bad data to
the back end if they renege. Still need to get the isAuthenticated
functionality running, then deploy. Once this is done, I'll be good.