Able to get functionality working as desired in my-account
Was actually able to get the display of the my-account route to alter
varying on whether there was a user or not. Next step is probably to
do this for the donate form, which is registering is authenticated as
false regardless of whether or not there is a user. This is probably
due to the fact that this is a nested component, will work to fix