Able to get proper request URL, cannot route to it
Need to figure out how I would go about sending my front end to the
proper authentication page; I am able to do get the request URL, but I
am not sure how to get to it with ember. Was not a problem in the rails
examples because I was using rails views.