Having some trouble redirecting client
I am able to make the request from my front end, but it ends up hanging
because emebr has no idea how to properly route from the front end at
the same time as routing from the back end. This is a fairly complicated
problem; I honestly Am unsure of how to solve it, largely because I am
nearly positive that the generation of the url leading to the redirect
page is closer to a side effect than an intended result. I need to find
some way to take the correct URL and plug it into the  window.location.replace()
method, but I have no idea how I am supposed to do this.